+++ src/TaskRunner.js
@@ -4,6 +4,7 @@ import cacache from 'cacache';
 import findCacheDir from 'find-cache-dir';
 import workerFarm from 'worker-farm';
 import serialize from 'serialize-javascript';
+import isWsl from 'is-wsl';
 
 import minify from './minify';
 
@@ -13,14 +14,19 @@ export default class TaskRunner {
   constructor(options = {}) {
     const { cache, parallel } = options;
     this.cacheDir =
-      cache === true ? findCacheDir({ name: 'terser-webpack-plugin' }) : cache;
+      cache === true
+        ? findCacheDir({ name: 'terser-webpack-plugin' }) || os.tmpdir()
+        : cache;
     // In some cases cpus() returns undefined
     // https://github.com/nodejs/node/issues/19022
     const cpus = os.cpus() || { length: 1 };
-    this.maxConcurrentWorkers =
-      parallel === true
-        ? cpus.length - 1
-        : Math.min(Number(parallel) || 0, cpus.length - 1);
+    // WSL sometimes freezes, error seems to be on the WSL side
+    // https://github.com/webpack-contrib/terser-webpack-plugin/issues/21
+    this.maxConcurrentWorkers = isWsl
+      ? 1
+      : parallel === true
+      ? cpus.length - 1
+      : Math.min(Number(parallel) || 0, cpus.length - 1);
   }
 
   run(tasks, callback) {